+++ packages/input-autocomplete/src/index.ts
@@ -1,5 +1,3 @@
-export {
-    InputAutocompleteDesktop as InputAutocomplete,
-    InputAutocompleteDesktopProps as InputAutocompleteProps,
-} from './Component.desktop';
+export { InputAutocompleteDesktop as InputAutocomplete } from './Component.desktop';
+export type { InputAutocompleteDesktopProps as InputAutocompleteProps } from './Component.desktop';
 export * from './responsive';