+++ packages/tooltip/src/index.ts
@@ -1,2 +1,3 @@
-export { TooltipDesktop as Tooltip, TooltipDesktopProps as TooltipProps } from './desktop';
+export { TooltipDesktop as Tooltip } from './desktop';
+export type { TooltipDesktopProps as TooltipProps } from './desktop';
 export * from './component.responsive';