+++ packages/masked-input/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 6.1.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@11.1.2
+
 ## 6.1.1
 
 ### Patch Changes