+++ packages/picker-button/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 10.2.1
+
+### Patch Changes
+
+### [#422](https://github.com/core-ds/core-components/pull/422)
+
+-   "export" типов заменен на "export type"
+
+-   Обновлены зависимости
+    -   select@13.2.4
+
 ## 10.2.0
 
 ### Minor Changes